Fix build
git-svn-id: https://svn.ices.utexas.edu/repos/iss/projects/GaloisCpp/trunk@11238 522c16a8-e32f-0410-b4cc-8071c0b41197

+++ test/sort.cpp
@@ -3,6 +3,7 @@
 
 #include <iostream>
 #include <cstdlib>
+#include <numeric>
 
 int RandomNumber () { return (rand()%1000000); }
 bool IsOdd (int i) { return ((i%2)==1); }
@@ -46,7 +47,7 @@ int do_sort() {
       std::sort(R.begin(), R.end());
       if (!std::equal(C.begin(), C.end(), R.begin()))
 	std::cout << "Cannot be made equal, sort mutated array\n";
-      for (int x = 0; x < V.size() ; ++x) {
+      for (size_t x = 0; x < V.size() ; ++x) {
 	std::cout << x << "\t" << V[x] << "\t" << C[x];
 	if (V[x] != C[x]) std::cout << "\tDiff";
 	if (V[x] < C[x]) std::cout << "\tLT";
